# src/service-account/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Микросервис для управления счетами и балансами.",
        "title": "Banking System - Account Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/api/accounts/{id}/transactions": {
            "get": {
                "description": "Возвращает список транзакций по счету с поддержкой пагинации",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Получить историю транзакций",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Public ID счета",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Количество записей (по умолчанию 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение (по умолчанию 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Начальная дата (RFC3339, например: 2026-02-01T00:00:00Z)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Конечная дата (RFC3339)",
                        "name": "end_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_Adopten123_banking-system_service-account_internal_domain.TransactionHistory"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Счет не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/accounts/{id}/transfer": {
            "post": {
                "description": "Осуществляет безопасный перевод денег между двумя счетами",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Перевод средств",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Public ID счета отправителя (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Уникальный ключ запроса",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Данные для перевода",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_http.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный перевод",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Счет не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/accounts/{id}/withdraw": {
            "post": {
                "description": "Списывает указанную сумму со счета. Учитывает кредитный лимит.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Снятие наличных",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID счета (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ключ идемпотентности (UUID)",
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Сумма для снятия",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Adopten123_banking-system_service-account_internal_domain.WithdrawRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное снятие (возвращает чек)",
                        "schema": {
                            "$ref": "#/definitions/github_com_Adopten123_banking-system_service-account_internal_domain.WithdrawResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос, нехватка средств или счет неактивен",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Счет не найден",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/internal_handler_http.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_Adopten123_banking-system_service-account_internal_domain.TransactionHistory": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "category_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "currency_code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "status_id": {
                    "type": "integer"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
        "github_com_Adopten123_banking-system_service-account_internal_domain.WithdrawRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                }
            }
        },
        "github_com_Adopten123_banking-system_service-account_internal_domain.WithdrawResponse": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "new_balance": {
                    "type": "number"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "internal_handler_http.TransferRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "currency_code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "to_account_id": {
                    "type": "string"
                }
            }
        },
...
    }
}